fix request header alignment for kvm patch kvm-qemu-v2.8

+++ libvmi/driver/kvm/kvm.c
@@ -58,7 +58,7 @@
 
 // request struct matches a definition in qemu source code
 struct request {
-    uint8_t type;   // 0 quit, 1 read, 2 write, ... rest reserved
+    uint64_t type;   // 0 quit, 1 read, 2 write, ... rest reserved
     uint64_t address;   // address to read from OR write to
     uint64_t length;    // number of bytes to read OR write
 };
